requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
pandas>=2.1.0
python-dotenv>=1.0.0

# Browser automation
webdriver-manager>=4.0.0
playwright>=1.40.0

# Data processing
openpyxl>=3.1.0
python-docx>=1.1.0
PyPDF2>=3.0.0

# Email
email-validator>=2.1.0

# Optional: AI features
# openai>=1.3.0
# anthropic>=0.7.0

# Extraction Tasks

pyyaml>=6.0
beautifulsoup4>=4.11.0
requests>=2.28.0
lxml>=4.9.0
